实现AMSI业务侧主动拉取与DLL升级unload状态切换
1. 将amsi_detect_config调整为Host侧通知服务，DLL侧作为客户端等待通知后主动拉取规则状态。

2. 规则快照增加desiredRuntimeState/state/ruleVersion封装，DLL根据running/unloading状态决定恢复检测或进入惰性卸载。

3. 将DLL升级unload流程下沉到AmsiDetectDllManager::UpdateAmsiDll，只有hash不一致时才进入unloading并广播unload。

4. AmsiIpcRuntime增加升级前快照保存与失败恢复能力，避免替换失败后长期停留unloading。

5. 清理业务侧规则快照定义与Lua预编译迁移后的旧源码引用，并补充整数/状态处理调整。

+++ src/module_amsi_detect/include/AmsiGlobalConf.h
@@ -74,19 +74,19 @@ namespace Engine {
             return m_statusEnqueueTimeoutMs;
         }
 
-        int GetRulePipeNums() const {
+        uint32_t GetRulePipeNums() const {
             return m_rulePipeNums;
         }
 
-        int GetEventPipeThreads() const {
+        uint32_t GetEventPipeThreads() const {
             return m_eventPipeThreads;
         }
 
-        int GetStatusPipeThreads() const {
+        uint32_t GetStatusPipeThreads() const {
             return m_statusPipeThreads;
         }
 
-        uint32_t GetDiagDroppedSummaryIntervalMs() const {
+        size_t GetDiagDroppedSummaryIntervalMs() const {
             return m_diagDroppedSummaryIntervalMs;
         }
 
@@ -111,12 +111,12 @@ namespace Engine {
         uint32_t m_statusEnqueueTimeoutMs = 50;
 
         // IPC 管道线程数配置.
-        size_t m_rulePipeNums = 8;
-        size_t m_eventPipeThreads = 4;
-        size_t m_statusPipeThreads = 2;
+        uint32_t m_rulePipeNums = 8;
+        uint32_t m_eventPipeThreads = 4;
+        uint32_t m_statusPipeThreads = 2;
 
         // DLL 诊断日志 dropped 汇总打印间隔，默认 60 秒.
-        uint32_t m_diagDroppedSummaryIntervalMs = 60 * 1000;
+        size_t m_diagDroppedSummaryIntervalMs = 60 * 1000;
     };
 }
